Move to patchwork-extensions-item module
