+++ patchwork-extensions-item/build.gradle
@@ -1,0 +1,6 @@
+archivesBaseName = "patchwork-extensions-item"
+version = getSubprojectVersion(project, "0.1.0")
+
+dependencies {
+	compile project(path: ':patchwork-fml', configuration: 'dev')
+}
