+++ patchwork-extensions-item/src/main/java/com/patchworkmc/mixin/extensions/item/ItemTagsAccessor.java
@@ -17,19 +17,17 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  */
 
-package com.patchworkmc.mixin.extension;
+package com.patchworkmc.mixin.extensions.item;
 
 import org.spongepowered.asm.mixin.Mixin;
-import net.minecraftforge.common.extensions.IForgeItem;
+import org.spongepowered.asm.mixin.gen.Accessor;
 
-import net.minecraft.entity.LivingEntity;
-import net.minecraft.item.AxeItem;
-import net.minecraft.item.ItemStack;
+import net.minecraft.tag.ItemTags;
 
-@Mixin(AxeItem.class)
-public abstract class MixinAxeItem implements IForgeItem {
-	@Override
-	public boolean canDisableShield(ItemStack stack, ItemStack shield, LivingEntity entity, LivingEntity attacker) {
-		return true;
+@Mixin(ItemTags.class)
+public class ItemTagsAccessor {
+	@Accessor
+	public static int getLatestVersion() {
+		throw new UnsupportedOperationException();
 	}
 }
