+++ patchwork-extensions/build.gradle
@@ -1,6 +1,2 @@
 archivesBaseName = "patchwork-extensions"
 version = getSubprojectVersion(project, "0.1.0")
-
-dependencies {
-	compile project(path: ':patchwork-fml', configuration: 'dev')
-}
