+++ settings.gradle
@@ -19,6 +19,7 @@ include 'patchwork-events-lifecycle'
 include 'patchwork-events-world'
 include 'patchwork-extensions-shearing'
 include 'patchwork-extensions'
+include 'patchwork-extensions-item'
 include 'patchwork-fml'
 include 'patchwork-registries'
 include 'patchwork-gui'
